requirements.txt
# Web framework
Flask==3.1.1
MarkupSafe==3.0.2

# HTTP + parsing
requests==2.32.4
beautifulsoup4==4.13.4
markdown==3.4.1

# Google AI SDKs
google-adk==1.7.0
google-genai==1.26.0
google-cloud-aiplatform==1.104.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
